Updates
git-svn-id: https://keyvault-org.googlecode.com/svn/trunk@9 f5e68fdc-ee4c-11de-9b30-9f9d59cce923

+++ encryption.h
@@ -1,7 +1,9 @@
 #ifndef _encryption_h_
 #define _encryption_h_
 
-unsigned char* create_random_array(ssize_t len);
+#include <stdint.h>
+
+unsigned char* create_random_array(int len);
 void shuffle_ivec(unsigned char ivec[16]);
 void aes_ofb(void* buffer,size_t len,const char* passphrase,const unsigned char* source_ivec);
